requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.1
importlib-metadata==0.23
more-itertools==7.2.0
nose==1.3.7
packaging==19.2
Pillow==6.2.0
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytesseract==0.3.0
pytest==5.2.2
pytest-html==2.0.0
pytest-metadata==1.8.0
pytest-rerunfailures==7.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.6
wcwidth==0.1.7
xlrd==1.2.0
zipp==0.6.0
